Rename UnderlingController and move it to underling package
