+++ src/main/java/com/mraof/minestuck/entity/underling/UnderlingSpawnSettings.java
@@ -1,11 +1,10 @@
-package com.mraof.minestuck.skaianet;
+package com.mraof.minestuck.entity.underling;
 
 import com.mraof.minestuck.MinestuckConfig;
-import com.mraof.minestuck.entity.MSEntityTypes;
-import com.mraof.minestuck.entity.underling.UnderlingEntity;
-import com.mraof.minestuck.event.UnderlingSpawnListEvent;
 import com.mraof.minestuck.alchemy.GristHelper;
 import com.mraof.minestuck.api.alchemy.GristType;
+import com.mraof.minestuck.entity.MSEntityTypes;
+import com.mraof.minestuck.event.UnderlingSpawnListEvent;
 import com.mraof.minestuck.world.lands.GristLayerInfo;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Vec3i;
@@ -22,7 +21,7 @@ import java.util.List;
  * Currently generates grist types for underlings and determines the underling spawn weights based on the distance from spawn.
  * Will probably take on generation of prototyping properties once we have that.
  */
-public final class UnderlingController
+public final class UnderlingSpawnSettings
 {
 	public static GristType getUnderlingType(UnderlingEntity entity)
 	{
